+++ src/main/java/team/incude/gsmc/v2/global/security/config/SecurityConfig.java
@@ -10,7 +10,7 @@ import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import team.incude.gsmc.v2.global.security.jwt.auth.filter.JwtAuthenticationFilter;
-import team.incude.gsmc.v2.global.security.jwt.usecase.JwtParserUseCase;
+import team.incude.gsmc.v2.global.security.jwt.application.usecase.JwtParserUseCase;
 
 @Configuration
 @EnableWebSecurity
