+++ src/main/java/team/incude/gsmc/v2/global/security/jwt/application/port/JwtApplicationPort.java
@@ -1,0 +1,4 @@
+package team.incude.gsmc.v2.global.security.jwt.application.port;
+
+public interface JwtApplicationPort {
+}
