+++ src/main/java/team/incude/gsmc/v2/global/security/jwt/application/usecase/JwtRefreshManagementUseCase.java
@@ -1,4 +1,4 @@
-package team.incude.gsmc.v2.global.security.jwt.usecase;
+package team.incude.gsmc.v2.global.security.jwt.application.usecase;
 
 public interface JwtRefreshManagementUseCase {
     void deleteRefreshToken(String refreshToken);
