+++ src/main/java/team/incude/gsmc/v2/global/security/jwt/application/usecase/service/JwtIssueService.java
@@ -1,4 +1,4 @@
-package team.incude.gsmc.v2.global.security.jwt.usecase.service;
+package team.incude.gsmc.v2.global.security.jwt.application.usecase.service;
 
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
@@ -7,10 +7,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import team.incude.gsmc.v2.domain.member.domain.constant.MemberRole;
-import team.incude.gsmc.v2.global.security.jwt.usecase.JwtIssueUseCase;
+import team.incude.gsmc.v2.global.security.jwt.application.usecase.JwtIssueUseCase;
 import team.incude.gsmc.v2.global.security.jwt.persistence.entity.RefreshTokenRedisEntity;
 import team.incude.gsmc.v2.global.security.jwt.persistence.repository.RefreshTokenRedisRepository;
-import team.incude.gsmc.v2.global.security.jwt.dto.TokenDto;
+import team.incude.gsmc.v2.global.security.jwt.presentation.dto.TokenDto;
 
 import javax.crypto.SecretKey;
 import java.time.LocalDateTime;
