+++ src/test/java/team/incude/gsmc/v2/domain/auth/application/usecase/service/RefreshServiceTest.java
@@ -12,10 +12,10 @@ import team.incude.gsmc.v2.domain.auth.presentation.data.response.AuthTokenRespo
 import team.incude.gsmc.v2.domain.member.application.port.MemberPersistencePort;
 import team.incude.gsmc.v2.domain.member.domain.Member;
 import team.incude.gsmc.v2.domain.member.domain.constant.MemberRole;
-import team.incude.gsmc.v2.global.security.jwt.dto.TokenDto;
-import team.incude.gsmc.v2.global.security.jwt.usecase.service.JwtIssueService;
-import team.incude.gsmc.v2.global.security.jwt.usecase.service.JwtParserService;
-import team.incude.gsmc.v2.global.security.jwt.usecase.service.JwtRefreshManagementService;
+import team.incude.gsmc.v2.global.security.jwt.presentation.dto.TokenDto;
+import team.incude.gsmc.v2.global.security.jwt.application.usecase.service.JwtIssueService;
+import team.incude.gsmc.v2.global.security.jwt.application.usecase.service.JwtParserService;
+import team.incude.gsmc.v2.global.security.jwt.application.usecase.service.JwtRefreshManagementService;
 
 import java.time.LocalDateTime;
