+++ src/test/java/team/incude/gsmc/v2/domain/certificate/application/usecase/service/UpdateCurrentCertificateServiceTest.java
@@ -21,7 +21,7 @@ import team.incude.gsmc.v2.domain.evidence.domain.constant.ReviewStatus;
 import team.incude.gsmc.v2.domain.member.application.port.MemberPersistencePort;
 import team.incude.gsmc.v2.domain.member.domain.Member;
 import team.incude.gsmc.v2.domain.score.domain.Score;
-import team.incude.gsmc.v2.global.security.jwt.usecase.service.CurrentMemberProvider;
+import team.incude.gsmc.v2.global.security.jwt.application.usecase.service.CurrentMemberProvider;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
