+++ src/test/java/team/incude/gsmc/v2/domain/evidence/application/usecase/service/CreateOtherScoringEvidenceServiceTest.java
@@ -23,7 +23,7 @@ import team.incude.gsmc.v2.domain.score.application.port.ScorePersistencePort;
 import team.incude.gsmc.v2.domain.score.domain.Category;
 import team.incude.gsmc.v2.domain.score.domain.Score;
 import team.incude.gsmc.v2.global.event.ScoreUpdatedEvent;
-import team.incude.gsmc.v2.global.security.jwt.usecase.service.CurrentMemberProvider;
+import team.incude.gsmc.v2.global.security.jwt.application.usecase.service.CurrentMemberProvider;
 
 import java.io.IOException;
 import java.util.concurrent.CompletableFuture;
